+++ internal/acctest/oauthissuer/oauth_issuer_test.go
@@ -75,6 +75,9 @@ resource "pingfederate_oauth_issuer" "%[1]s" {
   host        = "%[4]s"
   name        = "%[5]s"
   path        = "%[6]s"
+}
+data "pingfederate_oauth_issuer" "%[1]s" {
+  issuer_id = pingfederate_oauth_issuer.%[1]s.id
 }`, resourceName,
 		resourceModel.stateId,
 		resourceModel.description,