+++ internal/acctest/serversettingsgeneralsettings/server_settings_general_settings_test.go
@@ -80,6 +80,9 @@ func testAccServerSettingsGeneralSettings(resourceName string, resourceModel *se
 	return fmt.Sprintf(`
 resource "pingfederate_server_settings_general_settings" "%s" {
 	%s
+}
+data "pingfederate_server_settings_general_settings" "%[1]s" {
+  depends_on = [pingfederate_server_settings_general_settings.%[1]s]
 }`, resourceName,
 		optionalHcl,
 	)